refactor(credit): one rule for turning a reported dollar figure into money
The same conversion existed three times with three different sets of
guards: the credit context clamped at the ceiling and refused NaN, the
creation path refused NaN but did not clamp, and the run path guarded
nothing at all. Nothing was mis-billed — the out-of-range values were
caught downstream and reported as uncharged — but which of the three was
the rule could only be answered by reading all three.

The two credit ports now speak dollars, which is what the gateway reports
and what the caller actually has. The conversion belongs to the context
that owns money, so entrypoint/wiring does it at the boundary, where it can
already see both sides; no cross-context import and no depguard change.
Both local copies are gone.

The single surviving copy was also wrong in a way all three shared: it
scaled to micros and then converted to int64. A dollar figure large enough
to put that product past int64 converts implementation-defined and lands
back inside the billable range — 1e300 became -9223372036854775808, a
negative charge that the downstream guard would have refused for the wrong
reason. The comparison now happens in float space, before the conversion.

Proof: restoring the convert-then-compare order turns
TestAFigureTooLargeToScaleIsCappedNotWrapped red on both large rows.

+++ apps/platform/internal/entrypoint/api/apiserver/run_orchestration_integration_test.go
@@ -1628,7 +1628,7 @@ func TestARunWhoseCostCouldNotBeChargedIsNotReportedAsCleanedUp(t *testing.T) {
 	svc.Store = a.packages
 	svc.PollInterval = 20 * time.Millisecond
 	settled := 0
-	svc.CreditSettle = func(context.Context, pgx.Tx, pgtype.UUID, pgtype.UUID, *int64, int64) error {
+	svc.CreditSettle = func(context.Context, pgx.Tx, pgtype.UUID, pgtype.UUID, *float64, float64) error {
 		settled++
 		if settled == 1 {
 			return errors.New("the ledger is not answering")
